chore: lw7 pre release ready
